Updates to a custom stm32g4 dependency.
The new stm32g4 dependency is based on the `12.0` version which had some api
changes compared to the previously used "8.0" version.
This commit thus leaves the repository in an uncompilable state until the api changes are corrected.

Further cleans the features section of the manifest file to be updated with the
features provided by the current version of the crate.