Updates to the 12.0 api.
The newly introduced custom `stm32g4` crate dependency is based on an more
recent version of `stm32-rs` than the one that was previously used.

Some breaking api changes were introduced between the two versions such that the
update broke the code.

In particular:

- The `pllsys*` named items were renamed to `pll*`
- The `lscc*` named items were renamed to `lsc*`
- The `rc48*` named items were renamed to `hsi48*`
- `ccipr1` was renamed to `ccipr`

The code was thus updated to the new conventions.

The code was only tested with the currently active features.

Furthermore, rustfmt was run on the touched files.